requirements.txt
Django==1.8.7
djangorestframework==3.3.2
gunicorn==19.4.1
wheel==0.24.0
django-model-utils==2.0.3
Pillow==2.5.0
redis
django-ipware==1.1.3
django-cors-headers
gunicorn==19.4.5

# Only needed for running tests.
tox
